+++ packages/components-theme-editor/src/ThemeEditor.tsx
@@ -24,26 +24,34 @@
 
  */
 
-import { Grid } from '@looker/components'
+import { IconButton, DialogManager } from '@looker/components'
 import React, { FC } from 'react'
-import { Editor } from './Editor.stories'
+import {
+  ThemeEditorContent,
+  ThemeEditorContentProps,
+} from './ThemeEditorContent'
+import { ThemeEditableProps } from './types'
 
-export const Compare: FC = () => (
-  <Grid m="xlarge" gap="large" columns={4}>
-    <Editor name="Default" />
-    <Editor name="Generated" keyColor="#6C43E0" />
-    <Editor name="Customer Blue" keyColor="#116DFF" />
-    <Editor
-      name="THUNDER Salmon"
-      keyColor="#ff3ca0"
-      background="#000000"
-      text="#FFFFFF"
-    />
-  </Grid>
-)
-
-export default {
-  title: 'Theme',
+export interface ThemeEditorProps extends ThemeEditorContentProps {
+  hasCustomTheme: boolean
+  updateTheme: (customTheme?: ThemeEditableProps) => void
 }
 
-export * from './Editor.stories'
+export const ThemeEditor: FC<ThemeEditorProps> = ({
+  hasCustomTheme,
+  updateTheme,
+}) => (
+  <DialogManager
+    content={<ThemeEditorContent updateTheme={updateTheme} />}
+    width="90%"
+    maxWidth="90%"
+  >
+    <IconButton
+      icon="Beaker"
+      label="Customize Theme"
+      tooltipDisabled
+      color={hasCustomTheme ? 'key' : 'neutral'}
+      size="small"
+    />
+  </DialogManager>
+)